Added splitter, put TMS models in separate file, N+3 TMS in progress
