+++ openconcept/components/__init__.py
@@ -2,5 +2,5 @@ from .battery import SimpleBattery
 from .generator import SimpleGenerator
 from .motor import SimpleMotor
 from .propeller import SimplePropeller
-from .splitter import PowerSplit
+from .splitter import PowerSplit, FlowSplit
 from .turboshaft import SimpleTurboshaft
